+++ src/test/utils/checkAnswersConstants.ts
@@ -3,14 +3,14 @@ export const CLAIM_ID = 'claimId';
 // -- Summary Section Array Index
 export const INDEX_DETAILS_SECTION = 0;
 export const INDEX_RESPONSE_CLAIM_SECTION = 2;
-export const INDEX_FINANCIAL_SECTION = 4;
+export const INDEX_FINANCIAL_SECTION = 5;
 
-export const INDEX_RESPONSE_CLAIM_SECTION_PART_ADMISSION = 3;
-export const INDEX_FINANCIAL_SECTION_PART_ADMISSION = 5;
+export const INDEX_RESPONSE_CLAIM_SECTION_PART_ADMISSION = 4;
+export const INDEX_FINANCIAL_SECTION_PART_ADMISSION = 6;
 
-export const INDEX_RESPONSE_DETAILS_SECTION = 1;
-export const INDEX_RESPONSE_PAYMENT_SECTION = 6;
-export const INDEX_RESPONSE_FREE_TELEPHONE_MEDIATION_SECTION = 7;
+export const INDEX_RESPONSE_DETAILS_SECTION = 3;
+export const INDEX_RESPONSE_PAYMENT_SECTION = 7;
+export const INDEX_RESPONSE_FREE_TELEPHONE_MEDIATION_SECTION = 8;
 
 // -- Section Titles
 export const PAGES_CHECK_YOUR_ANSWER_YOUR_FINANCIAL_DETAILS_TITLE = 'PAGES.CHECK_YOUR_ANSWER.YOUR_FINANCIAL_DETAILS_TITLE';
